test: - Move command-registration test to new file Commands.test.ts
